+++ README.md
@@ -1,26 +1,20 @@
 ## 制作アプリのタイトル
-森高千里カレンダーver.3
+森高千里カレンダーver.4
 
 ## 制作アプリの説明（40文字程度）
-過去の課題で制作したカレンダーをPHP&SQLによる構築に改修工事したもの。
+前回のカレンダーver.3に予定の編集更新、削除機能を追加したもの。
 
 ## 工夫した点・こだわった点
-- 当該カレンダーアプリは、カレンダーの描画とともに、スケジュールデータ登録・保存し表示する。
-初版ではデータをローカルストレージに保存、ver.2ではFirestore Databaseへ保存というプログラムで作成した。
-今回ver.3として、サーバのデータベースへの保存に作り替えたもの。
-- 繰り返し予定を作るループ処理をJS上からPHP上で行うようになった。1回の予定エントリーで複数個のデータ登録を行う。
-のちのちの拡張のために、ひとつの繰り返し予定をグループとしてまとめるユニークなidを生成し、付加。
-- read.phpで読み込んだデータをJSON形式で、axiosを使ってmain.jsに渡すようにした。
-- 関数をfuncs.phpにまとめた。接続に関しても関数にして、connect.phpに。
-- 新たに予定のキーワード検索を突貫工事で追加した。
+- 登録された予定を呼び出し、編集画面を表示して編集更新する。単純な予定の更新以外に、繰り返し予定として登録されたものは、編集画面にその日だけ変更するか、その日以降の予定をすべて変更するか、の選択肢を表示させ、条件分岐してデータを更新するようにした。
+- 削除機能も編集更新と同様に、繰り返し予定の場合は、その日のみ削除か、その日以降すべて削除か、条件分岐できるようにした。
+- 前回追加した機能「予定検索」で検索した結果に対しても、結果の予定をクリックすると、その画面から編集画面に飛んで編集更新、削除ができるようにした。
+- 結果的に、すべてひとつのモーダル画面を共用して、ボタン選択だけで予定の登録、更新、削除ができるようになった。
 
 ## 次回トライしたこと（または機能）
-- 編集、削除に関する機能。次の講義でやると思うもの。
-- ver.2では、Firestoreによるユーザー認証を実装していたが、この先の講義で出てくると思われ
-今回は外しているので、のちのちの講義のあとで実装していく想定。
+- ユーザー認証機能
+- 画像の登録や日記への発展なども検討したい
 
 ## 備考（感想、シェアしたいこと等なんでも）
-- PHPに入ってから、キャッシュをクリアしないとデータが更新されないことが多い。
-実際はエラーが解消されていても、キャッシュが残って見た目上、エラーが残り続けるので、
-エラーがあった場合は、スペルミスチェックとキャッシュクリアをまず疑うようにする。
-- サーバへのデプロイが順調にできたのでホッとしている。
+- すでに登録している予定の編集更新、削除は容易そうに思えるかもしれないが、結構難しい。繰り返しの予定の編集には大変苦労した。
+- main.jsから$.post()メソッドでupdate.phpにデータを送る仕様になっているが、update.phpでvar_dumpすると、$_POSTの中身が常に空になっている現象に悩まされ、ついにXAMPPを再インストールするに至るも解決せず、ほぼ1日の時間を大きく浪費してしまったのが誠に悔しい。
+
